requirements.txt
ping3==4.0.4
influxdb==5.3.1
PyYAML==6.0.1
python-dotenv==1.0.0
